cos/queue_service: extract service stage into queue_service/service.rs (#1035 P3)
Move the four service_exact_*_queue_direct fns (~620 prod LOC) into
a sibling service.rs:

- service_exact_local_queue_direct (FIFO local)
- service_exact_local_queue_direct_flow_fair (MQFQ local)
- service_exact_prepared_queue_direct (FIFO prepared)
- service_exact_prepared_queue_direct_flow_fair (MQFQ prepared)

These are the **service** stage of the queue-service pipeline:
`select → service → drain → release/restore → settle`. They
orchestrate the per-queue token-gate / drain / TX-submit / settle
handshake for the four (local|prepared) × (FIFO|MQFQ) combinations
of exact-queue servicing.

queue_service/mod.rs production LOC (excluding mod tests):
  2,558 → **1,951** — UNDER the 2,000 modularity-discipline
  threshold for the first time since this issue was filed.
queue_service/service.rs: 628 LOC.

build clean (cargo build --release), 865 Rust tests pass. Cluster
smoke required (touches the worker service hot path).
