+++ userspace-dp/src/afxdp/cos/queue_service/mod.rs
@@ -45,6 +45,14 @@ pub(in crate::afxdp) use drain::{
     drain_exact_prepared_fifo_items_to_scratch, drain_exact_prepared_items_to_scratch_flow_fair,
 };
 
+// #1035 P3: service stage of the queue service pipeline (the four
+// service_exact_*_queue_direct fns) split into a sibling submodule.
+mod service;
+use service::{
+    service_exact_local_queue_direct, service_exact_local_queue_direct_flow_fair,
+    service_exact_prepared_queue_direct, service_exact_prepared_queue_direct_flow_fair,
+};
+
 use super::tx_completion::{
     apply_cos_prepared_result, apply_cos_send_result,
     apply_direct_exact_send_result, cos_tick_for_ns,
@@ -609,624 +617,9 @@ pub(in crate::afxdp) fn select_cos_surplus_batch(root: &mut CoSInterfaceRuntime,
     None
 }
 
-#[inline]
-fn service_exact_local_queue_direct(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    secondary_budget: u64,
-    now_ns: u64,
-    shared_recycles: &mut Vec<(u32, u64)>,
-) -> bool {
-    let flow_fair = binding
-        .cos_interfaces
-        .get(&root_ifindex)
-        .and_then(|root| root.queues.get(queue_idx))
-        .map(|queue| queue.flow_fair)
-        .unwrap_or(false);
-    if flow_fair {
-        return service_exact_local_queue_direct_flow_fair(
-            binding,
-            root_ifindex,
-            queue_idx,
-            secondary_budget,
-            now_ns,
-            shared_recycles,
-        );
-    }
-    if binding.free_tx_frames.is_empty() {
-        let _ = reap_tx_completions(binding, shared_recycles);
-    }
-    let queue_dscp_rewrite = cos_queue_dscp_rewrite(binding, root_ifindex, queue_idx);
-    binding.scratch_exact_local_tx.clear();
-    let root_budget = binding
-        .cos_interfaces
-        .get(&root_ifindex)
-        .map(|root| root.tokens)
-        .unwrap_or(0);
-    let build = {
-        let root = match binding.cos_interfaces.get_mut(&root_ifindex) {
-            Some(root) => root,
-            None => return false,
-        };
-        let queue = match root.queues.get_mut(queue_idx) {
-            Some(queue) => queue,
-            None => return false,
-        };
-        drain_exact_local_fifo_items_to_scratch(
-            queue,
-            &mut binding.free_tx_frames,
-            &mut binding.scratch_exact_local_tx,
-            binding.umem.area(),
-            root_budget,
-            secondary_budget,
-            queue_dscp_rewrite,
-        )
-    };
-    match build {
-        ExactCoSScratchBuild::Ready => {}
-        ExactCoSScratchBuild::Drop {
-            error,
-            dropped_bytes,
-        } => {
-            release_exact_local_scratch_frames(
-                &mut binding.free_tx_frames,
-                &mut binding.scratch_exact_local_tx,
-            );
-            if dropped_bytes > 0 {
-                subtract_direct_cos_queue_bytes(binding, root_ifindex, queue_idx, dropped_bytes);
-            } else {
-                refresh_cos_interface_activity(binding, root_ifindex);
-            }
-            binding.live.tx_errors.fetch_add(1, Ordering::Relaxed);
-            // #710: the scratch-build fell through `ExactCoSScratchBuild::Drop`
-            // with a frame-level error (capacity or slice). Subset of
-            // tx_errors.
-            binding
-                .live
-                .tx_submit_error_drops
-                .fetch_add(1, Ordering::Relaxed);
-            binding.live.set_error(error);
-            return false;
-        }
-    }
-    if binding.scratch_exact_local_tx.is_empty() {
-        maybe_wake_tx(binding, true, now_ns);
-        binding
-            .live
-            .set_error("no free TX frame available".to_string());
-        return false;
-    }
 
-    let mut writer = binding
-        .tx
-        .transmit(binding.scratch_exact_local_tx.len() as u32);
-    let inserted = writer.insert(binding.scratch_exact_local_tx.iter().map(|req| XdpDesc {
-        addr: req.offset,
-        len: req.len,
-        options: 0,
-    }));
-    writer.commit();
-    drop(writer);
-    // #812 Codex round-1 HIGH #1: sample the submit stamp AFTER
-    // `writer.commit()` so a scheduler preemption between `insert`
-    // and the ring submit does NOT inflate the measured latency.
-    // Pre-commit stamping attributed the preemption window to the
-    // kernel (submit→completion), which is exactly the opposite of
-    // what we want to observe. A reused caller `now_ns` would still
-    // leak up to ~1 ms of worker-loop staleness, so we take a fresh
-    // `monotonic_nanos()` here rather than re-using one from the
-    // outer scope. Only the accepted prefix (`.take(inserted as
-    // usize)`) is stamped — the retry tail returns to
-    // `free_tx_frames` and MUST NOT be stamped.
-    let ts_submit = monotonic_nanos();
-    stamp_submits(
-        &mut binding.tx_submit_ns,
-        binding
-            .scratch_exact_local_tx
-            .iter()
-            .take(inserted as usize)
-            .map(|req| req.offset),
-        ts_submit,
-    );
 
-    if inserted == 0 {
-        let dropped = binding.scratch_exact_local_tx.len() as u64;
-        binding.dbg_tx_ring_full += 1;
-        count_tx_ring_full_submit_stall(binding, root_ifindex, queue_idx, dropped);
-        maybe_wake_tx(binding, true, now_ns);
-        release_exact_local_scratch_frames(
-            &mut binding.free_tx_frames,
-            &mut binding.scratch_exact_local_tx,
-        );
-        refresh_cos_interface_activity(binding, root_ifindex);
-        binding.live.set_error("tx ring insert failed".to_string());
-        return false;
-    }
-    binding.dbg_tx_ring_submitted += inserted as u64;
-    binding.outstanding_tx = binding.outstanding_tx.saturating_add(inserted);
 
-    let (sent_packets, sent_bytes) = settle_exact_local_fifo_submission(
-        binding
-            .cos_interfaces
-            .get_mut(&root_ifindex)
-            .and_then(|root| root.queues.get_mut(queue_idx)),
-        &mut binding.free_tx_frames,
-        &mut binding.scratch_exact_local_tx,
-        inserted as usize,
-    );
-    // #940: post-settle V_min publish. FIFO queues currently have
-    // vtime_floor=None so this is a no-op; kept for uniformity and
-    // to shield future flow_fair-FIFO adoption.
-    publish_committed_queue_vtime(
-        binding
-            .cos_interfaces
-            .get(&root_ifindex)
-            .and_then(|root| root.queues.get(queue_idx)),
-    );
-    apply_direct_exact_send_result(binding, root_ifindex, queue_idx, sent_packets, sent_bytes);
-    maybe_wake_tx(binding, true, now_ns);
-    sent_packets > 0 || sent_bytes > 0
-}
-
-#[inline]
-fn service_exact_local_queue_direct_flow_fair(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    secondary_budget: u64,
-    now_ns: u64,
-    shared_recycles: &mut Vec<(u32, u64)>,
-) -> bool {
-    if binding.free_tx_frames.is_empty() {
-        let _ = reap_tx_completions(binding, shared_recycles);
-    }
-    let queue_dscp_rewrite = cos_queue_dscp_rewrite(binding, root_ifindex, queue_idx);
-    binding.scratch_local_tx.clear();
-    let root_budget = binding
-        .cos_interfaces
-        .get(&root_ifindex)
-        .map(|root| root.tokens)
-        .unwrap_or(0);
-    let build = {
-        let root = match binding.cos_interfaces.get_mut(&root_ifindex) {
-            Some(root) => root,
-            None => return false,
-        };
-        let queue = match root.queues.get_mut(queue_idx) {
-            Some(queue) => queue,
-            None => return false,
-        };
-        drain_exact_local_items_to_scratch_flow_fair(
-            queue,
-            &mut binding.free_tx_frames,
-            &mut binding.scratch_local_tx,
-            binding.umem.area(),
-            root_budget,
-            secondary_budget,
-            queue_dscp_rewrite,
-        )
-    };
-    match build {
-        ExactCoSScratchBuild::Ready => {}
-        ExactCoSScratchBuild::Drop {
-            error,
-            dropped_bytes,
-        } => {
-            restore_exact_local_scratch_to_queue_head_flow_fair(
-                binding
-                    .cos_interfaces
-                    .get_mut(&root_ifindex)
-                    .and_then(|root| root.queues.get_mut(queue_idx)),
-                &mut binding.free_tx_frames,
-                &mut binding.scratch_local_tx,
-            );
-            if dropped_bytes > 0 {
-                subtract_direct_cos_queue_bytes(binding, root_ifindex, queue_idx, dropped_bytes);
-            } else {
-                refresh_cos_interface_activity(binding, root_ifindex);
-            }
-            binding.live.tx_errors.fetch_add(1, Ordering::Relaxed);
-            // #710: the scratch-build fell through `ExactCoSScratchBuild::Drop`
-            // with a frame-level error (capacity or slice). Subset of
-            // tx_errors.
-            binding
-                .live
-                .tx_submit_error_drops
-                .fetch_add(1, Ordering::Relaxed);
-            binding.live.set_error(error);
-            return false;
-        }
-    }
-    if binding.scratch_local_tx.is_empty() {
-        maybe_wake_tx(binding, true, now_ns);
-        binding
-            .live
-            .set_error("no free TX frame available".to_string());
-        return false;
-    }
-
-    let mut writer = binding.tx.transmit(binding.scratch_local_tx.len() as u32);
-    let inserted = writer.insert(
-        binding
-            .scratch_local_tx
-            .iter()
-            .map(|(offset, req)| XdpDesc {
-                addr: *offset,
-                len: req.bytes.len() as u32,
-                options: 0,
-            }),
-    );
-    writer.commit();
-    drop(writer);
-    // #812 Codex round-1 HIGH #1: submit stamp AFTER commit — see plan
-    // §3.1 submit-site table (this is the
-    // service_exact_local_queue_direct_flow_fair variant). Stamping
-    // post-commit prevents a preemption window between `insert` and
-    // ring submit from being attributed to submit→completion latency.
-    let ts_submit = monotonic_nanos();
-    stamp_submits(
-        &mut binding.tx_submit_ns,
-        binding
-            .scratch_local_tx
-            .iter()
-            .take(inserted as usize)
-            .map(|(offset, _)| *offset),
-        ts_submit,
-    );
-
-    if inserted == 0 {
-        let dropped = binding.scratch_local_tx.len() as u64;
-        binding.dbg_tx_ring_full += 1;
-        count_tx_ring_full_submit_stall(binding, root_ifindex, queue_idx, dropped);
-        maybe_wake_tx(binding, true, now_ns);
-        restore_exact_local_scratch_to_queue_head_flow_fair(
-            binding
-                .cos_interfaces
-                .get_mut(&root_ifindex)
-                .and_then(|root| root.queues.get_mut(queue_idx)),
-            &mut binding.free_tx_frames,
-            &mut binding.scratch_local_tx,
-        );
-        refresh_cos_interface_activity(binding, root_ifindex);
-        binding.live.set_error("tx ring insert failed".to_string());
-        return false;
-    }
-    binding.dbg_tx_ring_submitted += inserted as u64;
-    binding.outstanding_tx = binding.outstanding_tx.saturating_add(inserted);
-
-    let (sent_packets, sent_bytes) = settle_exact_local_scratch_submission_flow_fair(
-        binding
-            .cos_interfaces
-            .get_mut(&root_ifindex)
-            .and_then(|root| root.queues.get_mut(queue_idx)),
-        &mut binding.free_tx_frames,
-        &mut binding.scratch_local_tx,
-        inserted as usize,
-    );
-    // #940: post-settle V_min publish. Settle has already applied
-    // any partial-rollback push_fronts (which republished via the
-    // rollback hook), so queue.queue_vtime now reflects only the
-    // actually-shipped frames.
-    publish_committed_queue_vtime(
-        binding
-            .cos_interfaces
-            .get(&root_ifindex)
-            .and_then(|root| root.queues.get(queue_idx)),
-    );
-    apply_direct_exact_send_result(binding, root_ifindex, queue_idx, sent_packets, sent_bytes);
-    maybe_wake_tx(binding, true, now_ns);
-    sent_packets > 0 || sent_bytes > 0
-}
-
-#[inline]
-fn service_exact_prepared_queue_direct(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    secondary_budget: u64,
-    now_ns: u64,
-) -> bool {
-    let flow_fair = binding
-        .cos_interfaces
-        .get(&root_ifindex)
-        .and_then(|root| root.queues.get(queue_idx))
-        .map(|queue| queue.flow_fair)
-        .unwrap_or(false);
-    if flow_fair {
-        return service_exact_prepared_queue_direct_flow_fair(
-            binding,
-            root_ifindex,
-            queue_idx,
-            secondary_budget,
-            now_ns,
-        );
-    }
-    let queue_dscp_rewrite = cos_queue_dscp_rewrite(binding, root_ifindex, queue_idx);
-    binding.scratch_exact_prepared_tx.clear();
-    let root_budget = binding
-        .cos_interfaces
-        .get(&root_ifindex)
-        .map(|root| root.tokens)
-        .unwrap_or(0);
-    let build = {
-        let root = match binding.cos_interfaces.get_mut(&root_ifindex) {
-            Some(root) => root,
-            None => return false,
-        };
-        let queue = match root.queues.get_mut(queue_idx) {
-            Some(queue) => queue,
-            None => return false,
-        };
-        drain_exact_prepared_fifo_items_to_scratch(
-            queue,
-            &mut binding.scratch_exact_prepared_tx,
-            binding.umem.area(),
-            &mut binding.free_tx_frames,
-            &mut binding.pending_fill_frames,
-            binding.slot,
-            root_budget,
-            secondary_budget,
-            queue_dscp_rewrite,
-        )
-    };
-    match build {
-        ExactCoSScratchBuild::Ready => {}
-        ExactCoSScratchBuild::Drop {
-            error,
-            dropped_bytes,
-        } => {
-            release_exact_prepared_scratch(&mut binding.scratch_exact_prepared_tx);
-            if dropped_bytes > 0 {
-                subtract_direct_cos_queue_bytes(binding, root_ifindex, queue_idx, dropped_bytes);
-            } else {
-                refresh_cos_interface_activity(binding, root_ifindex);
-            }
-            binding.live.tx_errors.fetch_add(1, Ordering::Relaxed);
-            // #710: the scratch-build fell through `ExactCoSScratchBuild::Drop`
-            // with a frame-level error (capacity or slice). Subset of
-            // tx_errors.
-            binding
-                .live
-                .tx_submit_error_drops
-                .fetch_add(1, Ordering::Relaxed);
-            binding.live.set_error(error);
-            return false;
-        }
-    }
-    if binding.scratch_exact_prepared_tx.is_empty() {
-        return false;
-    }
-
-    if cfg!(feature = "debug-log") {
-        for req in &binding.scratch_exact_prepared_tx {
-            if let Some(frame_data) = binding
-                .umem
-                .area()
-                .slice(req.offset as usize, req.len as usize)
-            {
-                if frame_has_tcp_rst(frame_data) {
-                    binding.dbg_tx_tcp_rst += 1;
-                }
-            }
-        }
-    }
-
-    let mut writer = binding
-        .tx
-        .transmit(binding.scratch_exact_prepared_tx.len() as u32);
-    let inserted = writer.insert(binding.scratch_exact_prepared_tx.iter().map(|req| XdpDesc {
-        addr: req.offset,
-        len: req.len,
-        options: 0,
-    }));
-    writer.commit();
-    drop(writer);
-    // #812 Codex round-1 HIGH #1: submit stamp AFTER commit — plan
-    // §3.1 submit-site table (the service_exact_prepared_queue_direct
-    // variant). Post-commit stamping ensures the measurement reflects
-    // the moment the ring submission actually landed in the kernel,
-    // not the moment before a potential preemption window.
-    let ts_submit = monotonic_nanos();
-    stamp_submits(
-        &mut binding.tx_submit_ns,
-        binding
-            .scratch_exact_prepared_tx
-            .iter()
-            .take(inserted as usize)
-            .map(|req| req.offset),
-        ts_submit,
-    );
-
-    if inserted == 0 {
-        let dropped = binding.scratch_exact_prepared_tx.len() as u64;
-        binding.dbg_tx_ring_full += 1;
-        count_tx_ring_full_submit_stall(binding, root_ifindex, queue_idx, dropped);
-        maybe_wake_tx(binding, true, now_ns);
-        release_exact_prepared_scratch(&mut binding.scratch_exact_prepared_tx);
-        refresh_cos_interface_activity(binding, root_ifindex);
-        binding
-            .live
-            .set_error("prepared tx ring insert failed".to_string());
-        return false;
-    }
-    binding.dbg_tx_ring_submitted += inserted as u64;
-    binding.outstanding_tx = binding.outstanding_tx.saturating_add(inserted);
-
-    let (sent_packets, sent_bytes) = settle_exact_prepared_fifo_submission(
-        binding
-            .cos_interfaces
-            .get_mut(&root_ifindex)
-            .and_then(|root| root.queues.get_mut(queue_idx)),
-        &mut binding.scratch_exact_prepared_tx,
-        &mut binding.in_flight_prepared_recycles,
-        inserted as usize,
-    );
-    // #940: post-settle V_min publish. FIFO queues have
-    // vtime_floor=None today; no-op shield for future adoption.
-    publish_committed_queue_vtime(
-        binding
-            .cos_interfaces
-            .get(&root_ifindex)
-            .and_then(|root| root.queues.get(queue_idx)),
-    );
-    apply_direct_exact_send_result(binding, root_ifindex, queue_idx, sent_packets, sent_bytes);
-    maybe_wake_tx(binding, true, now_ns);
-    sent_packets > 0 || sent_bytes > 0
-}
-
-#[inline]
-fn service_exact_prepared_queue_direct_flow_fair(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    secondary_budget: u64,
-    now_ns: u64,
-) -> bool {
-    let queue_dscp_rewrite = cos_queue_dscp_rewrite(binding, root_ifindex, queue_idx);
-    binding.scratch_prepared_tx.clear();
-    let root_budget = binding
-        .cos_interfaces
-        .get(&root_ifindex)
-        .map(|root| root.tokens)
-        .unwrap_or(0);
-    let build = {
-        let root = match binding.cos_interfaces.get_mut(&root_ifindex) {
-            Some(root) => root,
-            None => return false,
-        };
-        let queue = match root.queues.get_mut(queue_idx) {
-            Some(queue) => queue,
-            None => return false,
-        };
-        drain_exact_prepared_items_to_scratch_flow_fair(
-            queue,
-            &mut binding.scratch_prepared_tx,
-            binding.umem.area(),
-            &mut binding.free_tx_frames,
-            &mut binding.pending_fill_frames,
-            binding.slot,
-            root_budget,
-            secondary_budget,
-            queue_dscp_rewrite,
-        )
-    };
-    match build {
-        ExactCoSScratchBuild::Ready => {}
-        ExactCoSScratchBuild::Drop {
-            error,
-            dropped_bytes,
-        } => {
-            restore_exact_prepared_scratch_to_queue_head_flow_fair(
-                binding
-                    .cos_interfaces
-                    .get_mut(&root_ifindex)
-                    .and_then(|root| root.queues.get_mut(queue_idx)),
-                &mut binding.scratch_prepared_tx,
-            );
-            if dropped_bytes > 0 {
-                subtract_direct_cos_queue_bytes(binding, root_ifindex, queue_idx, dropped_bytes);
-            } else {
-                refresh_cos_interface_activity(binding, root_ifindex);
-            }
-            binding.live.tx_errors.fetch_add(1, Ordering::Relaxed);
-            // #710: the scratch-build fell through `ExactCoSScratchBuild::Drop`
-            // with a frame-level error (capacity or slice). Subset of
-            // tx_errors.
-            binding
-                .live
-                .tx_submit_error_drops
-                .fetch_add(1, Ordering::Relaxed);
-            binding.live.set_error(error);
-            return false;
-        }
-    }
-    if binding.scratch_prepared_tx.is_empty() {
-        return false;
-    }
-
-    if cfg!(feature = "debug-log") {
-        for req in &binding.scratch_prepared_tx {
-            if let Some(frame_data) = binding
-                .umem
-                .area()
-                .slice(req.offset as usize, req.len as usize)
-            {
-                if frame_has_tcp_rst(frame_data) {
-                    binding.dbg_tx_tcp_rst += 1;
-                }
-            }
-        }
-    }
-
-    let mut writer = binding
-        .tx
-        .transmit(binding.scratch_prepared_tx.len() as u32);
-    let inserted = writer.insert(binding.scratch_prepared_tx.iter().map(|req| XdpDesc {
-        addr: req.offset,
-        len: req.len,
-        options: 0,
-    }));
-    writer.commit();
-    drop(writer);
-    // #812 Codex round-1 HIGH #1: submit stamp AFTER commit — plan
-    // §3.1 submit-site table (the
-    // service_exact_prepared_queue_direct_flow_fair variant). See the
-    // exact_local variant above for the preemption-window rationale.
-    let ts_submit = monotonic_nanos();
-    stamp_submits(
-        &mut binding.tx_submit_ns,
-        binding
-            .scratch_prepared_tx
-            .iter()
-            .take(inserted as usize)
-            .map(|req| req.offset),
-        ts_submit,
-    );
-
-    if inserted == 0 {
-        let dropped = binding.scratch_prepared_tx.len() as u64;
-        binding.dbg_tx_ring_full += 1;
-        count_tx_ring_full_submit_stall(binding, root_ifindex, queue_idx, dropped);
-        maybe_wake_tx(binding, true, now_ns);
-        restore_exact_prepared_scratch_to_queue_head_flow_fair(
-            binding
-                .cos_interfaces
-                .get_mut(&root_ifindex)
-                .and_then(|root| root.queues.get_mut(queue_idx)),
-            &mut binding.scratch_prepared_tx,
-        );
-        refresh_cos_interface_activity(binding, root_ifindex);
-        binding
-            .live
-            .set_error("prepared tx ring insert failed".to_string());
-        return false;
-    }
-    binding.dbg_tx_ring_submitted += inserted as u64;
-    binding.outstanding_tx = binding.outstanding_tx.saturating_add(inserted);
-
-    let (sent_packets, sent_bytes) = settle_exact_prepared_scratch_submission_flow_fair(
-        binding
-            .cos_interfaces
-            .get_mut(&root_ifindex)
-            .and_then(|root| root.queues.get_mut(queue_idx)),
-        &mut binding.scratch_prepared_tx,
-        &mut binding.in_flight_prepared_recycles,
-        inserted as usize,
-    );
-    // #940: post-settle V_min publish. Settle has applied any
-    // partial-rollback push_fronts via the rollback hook;
-    // queue.queue_vtime now reflects only actually-shipped frames.
-    publish_committed_queue_vtime(
-        binding
-            .cos_interfaces
-            .get(&root_ifindex)
-            .and_then(|root| root.queues.get(queue_idx)),
-    );
-    apply_direct_exact_send_result(binding, root_ifindex, queue_idx, sent_packets, sent_bytes);
-    maybe_wake_tx(binding, true, now_ns);
-    sent_packets > 0 || sent_bytes > 0
-}
